move genesis block hash comparison to unit test
